Add a heavy prime class

+++ src/cpp/math/method/prime/millerrabin.hpp
@@ -15,7 +15,7 @@
 //
 // %require
 // ```
-using ll = long long;
+using ll = int64_t;
 // ```
 //
 // %verified
